Move pkg/gzipfs to internal/gzipfs
