+++ internal/gzipfs/testdata/foobar
