+++ internal/gzipfs/testdata/foobar.gz
