+++ internal/gzipfs/testdata/foobaz.gz
