+++ internal/gzipfs/testdata/foobaz_truncated.gz
