+++ internal/gzipfs/testdata/not_archive.gz
